Reduce the complexity of the IsPrimeServer
I'm happier with this version but I'm not entirely sold on my use of
`with` in the core `respond_to_lines_until_closed` path.

One the one hand: it does make the happy path very clear at a high level
of indentation.

But on the other hand the mixing of errors from :gen_tcp and
Jason.decode make it hard to know which step in the with declaration had
the issue.

In many cases that wouldn't be a problem but in this case we need to
send a malformed response for some errors but not all "errors". A big
example is {:error, :closed} which simply means the client closed the
socket which is entirely expected.

The alternative would be a case statement for :gen_tcp.recv and then a
with statement or another case statement to handle Jason.decode. That
may be more readable as the levels of error handling responsibility
would be clearer.

Error from :gen_tcp.recv? Don't send a malformed response.

Error or missed pattern from Jason.decode? Send a malformed response.

Hrmmm I'll leave it as with for now since I really do like the clear
declaration of the happy path.