Increase buffer size for packet lines
Hey an edge my previous complex code handled.

Turns out the protohackers challenge actually hit the default buffer
limit for :gen_tcp.listen

```
dbg(:inet.getopts(listen_socket, [:buffer])) # => {:ok, [buffer: 1460]}
```

Later in the challenges

```
%Jason.DecodeError{position: 1460, token: nil, data: "{\"metho...
```

That's right, Protohackers had a valid JSON message that happened to
exceed the default buffer. That means the valid JSON was truncated and
Jason was unable to decode and we were incorrectly sending a malformed
response.

Fix that by increasing the buffer to a bigger but still reasonable
amount of 100KB.