Add UnusualDatabaseProtocolServer
Hey UDP! Fast, connection-less, order-less, state-less, error-prone
packets!

Ok maybe error-prone is too harsh, but I ran into errors with the
checks for this problem in which the Protohackers checks would
stop communicating with my server without any reason. I could still
query my server from other sources but the Protohackers checks just
didn't see any data.

The problem there could lie with my server, my fly.io server
configuration, fly.io's networking, the networking layers between
fly.io and Protohackers, Protohackers networking config, or the
Protohackers client. Or even a bunch of other places whee.

At any rate this was a very straightforward problem from the
application side. A little message splitting, a bit of data work,
no problem.

I took the opportunity to try out a server setup where I have an
`active: true` configuration. That means that instead of explicitly
receiving messages by calling `:gen_udp.recv` (or `:gen_tcp.recv`)
my server process will get the incoming network messages as Elixir
process messages.

I also leaned more into messages for handling the request/response
lifecycle for my GenServer instead of handling all of the interaction
at once.

How's that work? Let's walk through a couple requests.

The general flow

1. An incoming insert packet is handled by the `handle_info` function

(`handle_info` is the callback function that receives all Elixir
messages sent to the process)

2. That `handle_info` function recognizes a UDP packet and pulls apart
the relevant networking details. Then it determines if the message
is an insert or retrieve and sends itself a message with the relevant
details necessary to complete that work.

Actually that's not quite true, instead of a low level Elixir message
I use `GenServer.cast` to asynchronously send the message at a
slightly higher level of abstraction. The effect is pretty much the
same.

3. The specific `handle_cast` function to apply is patterned matched.
The `handle_cast` function does its specific work and updates the
GenServer state.

All GenServer callbacks return the new GenServer state as part of
their contract. That state may be changed or unchanged depending on
the specific work of the function. For an insert's case the state is
changed with new database data structure after the insert has been
applied. For a retrieve's case the state is unchanged.

INSERT

1. An incoming insert packet is handled by the `handle_info` function

2. That `handle_info` function recognizes a UDP packet and pulls apart
the relevant networking details. Then it determines the message is
an insert and calls `GenServer.cast` to send an insert message with
the details.

3. The specific `handle_cast` implementation for an insert is pattern
matched. That function is only concerned with updating the database in
the state.

RETRIEVE

1. An incoming retrieve packet is handled by the `handle_info`
function

2. That `handle_info` function recognizes a UDP packet and pulls apart
the relevant networking details. Then it determines the message is
a retrieve and calls `GenServer.cast` to send an insert message with
the details.

3. The specific `handle_cast` implementation for a retrieve is pattern
matched. That function is only concerned with querying the database
and sending a UDP packet back to the original message source.

VERSION

Nothing fancy here. "version" requests are specially recognized per
the spec and not even given a chance to incorrectly alter the database
because they get nowhere near the data handling.