Rewrite ChatRoomServer to use active sockets
Thanks to @whatyouhide for the insightful feedback pointing me in a
better direction!

The main effect of this change is to make the the input loop for a
connected user much more efficient.

As a quick ballpark my integration tests for the ChatRoomServer
previously took 2.2 seconds. On this new version they take 1.8 seconds.

Not a huge change to the clock but a significant improvement when you
think about the effects over time. The new active code is more
readable as well!

What changed?

My previous approach for the input loop was to bounce between
`:gen_tcp.receive` (for incoming chat messages) and `receive`
(for dispatched server messages). Both of those calls block and I
didn't want any clients to be blocked waiting for messages. So each
side of the loop had a short timeout of 100ms.

The loop:

* Try to receive a new TCP packet for 100ms
  * Got a packet? handle it as input
  * Timed out? Try to receive an Elixir message for 100ms
    * Got a message? handle it and restart the loop
    * Timed out? restart the loop

Now I've switched the TCP socket connections from using `active:
false` to `active: true`. That change means that TCP packets will be
sent to the socket controlling process as Elixir messages. THAT means
no more bouncing between two kinds of incoming data. Everything is now
an Elixir message and they can all be processed in one receive block
as they arrive. Nice!

A key function I had to find to make this work is
`:gen_tcp.controlling_process/2`. That function allows a process that
has accepted a TCP connection to hand off control to another process.
And control means that the new process will receive all the messages
for an active connection. Even queued messages for that socket that are
in the original processes mailbox will be migrated to the new process.

Very cool!