Add MobInTheMiddleServer
Protohackers problem 5: https://protohackers.com/problem/5

Ahh an actual machine-in-the-middle attack! And performed against our
own chat server from problem 3. That's cool!

The approach here is very straightforward now that I've learned how to
use active sockets.

Get a connection from a client, open a connection to the chat server,
and then whenever a message comes in from one side rewrite it and send
it to the other side.

Getting the regex right at the edges was more difficult than the
networking portion of this challenge. I resorted to cheating because
even I, a former Perl programmer, could not construct a regex that I
was happy with that also handled all the edges. So I gave up trying to
find the one regex and simply split all the messages by spaces, then
applied a simple regex against each word, then rejoined all the words
by spaces again. I'm not trimming any spaces so all whitespaces are
preserved. For sure more computationally expensive than a compiled
regex could be, but we're only talking chat messages here with maybe
a few dozen words at the most.

Another interesting aspect to this problem in this Elixir application
is that I'm really using the configuration now. By default the
MobInTheMiddleServer is configured to connect to localhost. In prod the
application is configured to connect to the defined problem server.