Solving Protohackers Problem 6
This commit solves
[Protohackers Problem 6](https://protohackers.com/problem/6)

The problem was a huge jump up in complexity from problems 1-5, but
that's great because I learned a lot!

This time around I opted to roll my own "database" instead of leaning
on Registry and wow I appreciate Registry even more now.

I went down a few false paths with this application architecture.

One of my earlier attempts was to prematurely setup camera and
dispatcher as two separate kinds of client. That worked great in one
aspect: I was able to very clearly define the messages that each kind
of client would accept.

But a major sticking point was handling heartbeats. Both cameras and
dispatchers can request a heartbeat which is handled the same for each
client.

I didn't want to spin off special heartbeat handling so that meant I
was duplicating the heartbeat code in both Camera and Dispatcher.

That approach fell down quickly in the protohacker tests. It turns out
that it's expected that any unspecified clients can request heartbeats:
they don't need to identify as a camera or dispatcher first. That means
I'd have to duplicate and specially handle heartbeats in three places
and handoff operational heartbeats that were started before a client
identified itself. Not great at all.

So I reworked into having a single client that stores the type of
client as part of its state. I was still able to express invalid
messages based on the type of client with guards but now it doesn't
read quite as clearly. But the benefit of having one flow for
heartbeats more than outweighs that drawback.

Other than figuring out all of the business logic edges a major issue I
ran into was fly.io default limits.

With the default connection limits I'd run into a "broken pipe" failure
for the test of a big world.

```
[5big.test] FAIL:Broken Pipe (server disconnected unexpectedly)
```

Thanks to Andrea's insight that was straightforward to fix by
increasing the concurrently limits in my fly config from 20/25 to
200/250.

The code I have in this commit works and passes the Protohacker tests,
but next I need to (re)learn how to setup proper supervision trees
below the application's supervisor because I'm starting too much from
one init function.