More OTP solution for Protohackers Problem 5
Thanks to @whatyouhide for his
[Protohackers in Elixir: day 5](https://youtu.be/tm4Jgg7zeXk) video!

His guide was a hugely great refresher for me into how to setup a
proper supervision tree including how to use `DynamicSupervisor` which
is new to me. The last time I did dynamic supervision the correct
approach was the, now deprecated, `:simple_one_for_one` strategy.

While my previous approach
[Protohackers.MobInTheMiddleServer](https://github.com/sdball/protohackers/blob/main/lib/protohackers/mob_in_the_middle_server.ex)
worked fine and was reasonably readable, it was overworking the Task
module and doing too much supervision from within a non-supervisor
process. It all worked because Elixir is awesome but it wasn't very
idiomatic and would not have tolerated failure as well as a proper OTP
solution.

With this redesign we have an actual supervision tree watching the
running "mob in the middle" server and connections.

We have a top level supervisor which runs an acceptor and connection
supervisor. Then the acceptor accepts a new socket connection it asks
the connection supervisor to start a connection process. The connection
process connects to the chat server being proxied. The connection
process is what actually handles the TCP socket traffic from the client
and the upstream server and does the actual work of rewriting the
boguscoin addresses per the problem requirements.

While this approach is somewhat more complex the actual implementation
details are easier. For example we don't need to handle our own
low-level `receive` block to handle the Elixir messages from the TCP
socket. We can instead rely on the GenServer behavior to deliver them
via the `handle_info` callback functions. Nice!